issue-89: initial implementation of search result page
