+++ frontend/src/pages/home/Home.tsx
@@ -1,19 +1,39 @@
-import { useEffect } from "react";
+import { useEffect, useRef, useState } from "react";
+import { Link, useHistory } from "react-router-dom";
 import { Timeline } from "react-twitter-widgets";
 import { useAppDispatch, useAppSelector } from "../../app/hooks";
+import { randomString } from "../../app/util";
 import Header from "../../components/Header";
-import HomeSearchBox from "./HomeSearchBox";
-import { getStats } from "./homeSlice";
+import Entity from "../../model/Entity";
+import Ontology from "../../model/Ontology";
+import Thing from "../../model/Thing";
+import { getSearchOptions, getStats } from "./homeSlice";
 
 export default function Home() {
   const dispatch = useAppDispatch();
+  const history = useHistory();
   const stats = useAppSelector((state) => state.home.stats);
+  const results = useAppSelector((state) => state.home.searchResult);
+  const loading = useAppSelector((state) => state.home.loadingSearchResult);
+  const homeSearch = document.getElementById("home-search") as HTMLInputElement;
 
-  // const [openList, setOpenList] = useState<boolean>(false);
+  const [open, setOpen] = useState<boolean>(false);
+  const [query, setQuery] = useState<string>("");
 
   useEffect(() => {
     dispatch(getStats());
   }, [dispatch]);
+  useEffect(() => {
+    dispatch(getSearchOptions(query));
+  }, [dispatch, query]);
+
+  const mounted = useRef(false);
+  useEffect(() => {
+    mounted.current = true;
+    return () => {
+      mounted.current = false;
+    };
+  });
 
   document.title = "Ontology Lookup Service (OLS)";
   return (
@@ -26,18 +46,114 @@ export default function Home() {
               <div className="text-3xl mb-4 text-neutral-black font-bold">
                 Welcome to the EMBL-EBI Ontology Lookup Service
               </div>
-              <div className="mb-4">
-                <HomeSearchBox />
+              <div className="flex flex-nowrap gap-4 mb-4">
+                <div className="relative w-full self-center">
+                  <input
+                    id="home-search"
+                    type="text"
+                    placeholder="Search OLS..."
+                    className="input-default text-lg focus:rounded-b-sm focus-visible:rounded-b-sm pl-3"
+                    onFocus={() => {
+                      setOpen(true);
+                    }}
+                    onBlur={() => {
+                      setTimeout(function () {
+                        if (mounted.current) setOpen(false);
+                      }, 500);
+                    }}
+                    value={query}
+                    onChange={(e) => {
+                      setQuery(e.target.value);
+                    }}
+                  />
+                  <div
+                    className={
+                      loading
+                        ? "spinner-default w-7 h-7 absolute right-3 top-2.5 z-10"
+                        : "hidden"
+                    }
+                  />
+                  <ul
+                    className={
+                      open
+                        ? "list-none bg-white text-neutral-dark border-2 border-neutral-dark shadow-input rounded-b-md w-full absolute left-0 top-12 z-10"
+                        : "hidden"
+                    }
+                  >
+                    {results.length === 0 ? (
+                      <div className="py-1 px-3 text-lg leading-loose">
+                        No options
+                      </div>
+                    ) : (
+                      results.map((option: Thing) => {
+                        const termUrl = encodeURIComponent(
+                          encodeURIComponent(option.getIri())
+                        );
+                        return (
+                          <li
+                            key={randomString()}
+                            className="py-2 px-3 hover:bg-link-light hover:rounded-sm hover:cursor-pointer"
+                          >
+                            {option instanceof Entity ? (
+                              <Link
+                                onClick={() => {
+                                  setOpen(false);
+                                }}
+                                to={`/ontologies/${option.getOntologyId()}/${option.getTypePlural()}/${termUrl}`}
+                              >
+                                <div className="flex justify-between">
+                                  <span className="truncate">
+                                    {option.getName()}
+                                  </span>
+                                  <span className="ml-2 text-right">
+                                    <span className="mr-2 truncate bg-link-default px-3 py-1 rounded-lg text-sm text-white uppercase">
+                                      {option.getOntologyId()}
+                                    </span>
+                                    <span className="bg-orange-default px-3 py-1 rounded-lg text-sm text-white uppercase">
+                                      {option.getShortForm()}
+                                    </span>
+                                  </span>
+                                </div>
+                              </Link>
+                            ) : null}
+                            {option instanceof Ontology ? (
+                              <Link
+                                onClick={() => {
+                                  setOpen(false);
+                                }}
+                                to={"/ontologies/" + option.getOntologyId()}
+                              >
+                                <div className="flex">
+                                  <span className="truncate text-link-dark font-bold">
+                                    {option.getName() || option.getOntologyId()}
+                                  </span>
+                                </div>
+                              </Link>
+                            ) : null}
+                          </li>
+                        );
+                      })
+                    )}
+                  </ul>
+                </div>
+                <button
+                  className="button-primary text-lg font-bold self-center"
+                  onClick={() => {
+                    history.push("/home/search/" + homeSearch.value);
+                  }}
+                >
+                  Search
+                </button>
               </div>
               <div className="grid grid-cols-2">
                 <div className="text-neutral-black">
                   <span>
                     Examples:&nbsp;
-                    <a href="#" className="link-default">
+                    <a href="home/search/diabetes" className="link-default">
                       diabetes
                     </a>
                     &#44;&nbsp;
-                    <a href="#" className="link-default">
+                    <a href="home/search/GO:0098743" className="link-default">
                       GO:0098743
                     </a>
                   </span>
